Rename folder to newsrec_pytorch
